Made AddressTranslation a common component. Added mocked send nft flow screens
