+++ source/components/AddressTranslation/components/AddressRow/styles.js
@@ -181,4 +181,34 @@ export default makeStyles((theme) => ({
     padding: '12px 10px 12px 0px',
     maxHeight: 130,
   },
+  leftContainer: {
+    maxWidth: 45,
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    padding: '8px 0',
+  },
+  rightContainer: {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'space-between',
+    width: '100%',
+    paddingTop: 6,
+  },
+  row: {
+    display: 'flex',
+    alignItems: 'center',
+  },
+  addressRow: {
+    display: 'flex',
+    alignItems: 'center',
+    width: '100%',
+    justifyContent: 'space-between',
+  },
+  relative: {
+    position: 'relative',
+  },
+  primaryAddressRow: {
+    paddingTop: 13,
+  },
 }));
